+++ tests/testthat/helper-options.R
@@ -1,7 +1,4 @@
 options(
-  # There is a warning when using Arrow about using pull. It says to set
-  # this option.
-  arrow.pull_as_vector = TRUE,
   # DuckDB outputs a lot of messages, this stops that.
   duckdb.materialize_message = FALSE
 )